requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
accelerate=1.7.0=pypi_0
anyio=3.7.1=pyhd8ed1ab_0
argon2-cffi=23.1.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310ha75aee5_5
asttokens=3.0.0=pyhd8ed1ab_1
attrs=24.2.0=pyh71513ae_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_4
backports.functools_lru_cache=2.0.0=pyhd8ed1ab_0
beautifulsoup4=4.12.3=pyha770c72_0
bleach=6.1.0=pyhd8ed1ab_0
bzip2=1.0.8=h4bc722e_7
ca-certificates=2025.4.26=hbd8a1cb_0
certifi=2025.4.26=pypi_0
cffi=1.17.1=py310h8deb56e_0
charset-normalizer=3.4.2=pypi_0
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.14=py310hf71b8c6_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
einops=0.8.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
filelock=3.18.0=pypi_0
fonttools=4.58.2=pypi_0
fsspec=2025.5.1=pypi_0
ftfy=6.3.1=pypi_0
hf-xet=1.1.3=pypi_0
huggingface-hub=0.32.4=pypi_0
idna=3.10=pyhd8ed1ab_0
importlib-metadata=8.6.1=pyha770c72_0
importlib_resources=6.0.0=pyhd8ed1ab_0
ipykernel=6.16.2=pyh210e3f2_0
ipython=8.21.0=pyh707e725_0
ipython_genutils=0.2.0=pyhd8ed1ab_1
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=pyhd8ed1ab_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter_client=7.4.9=pyhd8ed1ab_0
jupyter_core=5.8.1=pyh31011fe_0
jupyter_server=1.23.4=pyhd8ed1ab_0
jupyterlab_pygments=0.3.0=pyhd8ed1ab_1
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h659f571_0
ld_impl_linux-64=2.43=h712a8e2_4
libedit=3.1.20250104=pl5321h7949ede_0
libexpat=2.7.0=h5888daf_0
libffi=3.4.6=h2dba641_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgomp=15.1.0=h767d61c_2
liblzma=5.8.1=hb9d3cd8_1
liblzma-devel=5.8.1=hb9d3cd8_1
libnsl=2.0.1=hd590300_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.50.1=hee588c1_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=15.1.0=h4852527_2
libuuid=2.38.1=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libzlib=1.3.1=hb9d3cd8_2
markupsafe=3.0.2=py310h89163eb_1
matplotlib=3.10.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mistune=3.0.2=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
nbclassic=1.1.0=pyhd8ed1ab_0
nbclient=0.7.0=pyhd8ed1ab_0
nbconvert-core=7.6.0=pyhd8ed1ab_0
nbformat=5.8.0=pyhd8ed1ab_0
ncurses=6.5=h2d0b736_3
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.4.2=pypi_0
notebook=6.5.7=pyha770c72_0
notebook-shim=0.2.4=pyhd8ed1ab_0
numpy=2.2.6=pypi_0
nvidia-cublas-cu12=12.6.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.6.80=pypi_0
nvidia-cuda-nvrtc-cu12=12.6.77=pypi_0
nvidia-cuda-runtime-cu12=12.6.77=pypi_0
nvidia-cudnn-cu12=9.5.1.17=pypi_0
nvidia-cufft-cu12=11.3.0.4=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.7.77=pypi_0
nvidia-cusolver-cu12=11.7.1.2=pypi_0
nvidia-cusparse-cu12=12.5.4.2=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.26.2=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.6.77=pypi_0
open-clip-torch=2.32.0=pypi_0
openssl=3.5.0=h7b32b05_1
packaging=23.2=pyhd8ed1ab_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=11.2.1=pypi_0
pip=24.0=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_1
platformdirs=4.3.8=pyhe01879c_0
prometheus_client=0.17.1=pyhd8ed1ab_0
prompt-toolkit=3.0.48=pyha770c72_0
psutil=7.0.0=py310ha75aee5_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_1
pycparser=2.21=pyhd8ed1ab_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.2.3=pypi_0
pyrsistent=0.20.0=py310ha75aee5_1
python=3.10.18=hd6af730_0_cpython
python-dateutil=2.9.0=pyhd8ed1ab_0
python-fastjsonschema=2.20.0=pyhd8ed1ab_0
python_abi=3.10=7_cp310
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py310h71f11fc_1
readline=8.2=h8c095d6_2
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.5.3=pypi_0
send2trash=1.8.3=pyh0d859eb_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
smalldiffusion=0.4.3=pypi_0
sniffio=1.3.1=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
sqlite=3.50.1=h9eae976_0
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.14.0=pypi_0
terminado=0.17.1=pyh41d4057_0
timm=1.0.15=pypi_0
tinycss2=1.4.0=pyhd8ed1ab_0
tk=8.6.13=noxft_h4845f30_101
torch-ema=0.3=pypi_0
torchvision=0.22.1=pypi_0
tornado=6.5.1=py310ha75aee5_0
tqdm=4.67.1=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
triton=3.3.1=pypi_0
typing-extensions=4.14.0=pypi_0
tzdata=2025b=h78e105d_0
urllib3=2.4.0=pypi_0
wcwidth=0.2.10=pyhd8ed1ab_0
webencodings=0.5.1=pyhd8ed1ab_2
websocket-client=1.6.1=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
xz=5.8.1=hbcc6ac9_1
xz-gpl-tools=5.8.1=hbcc6ac9_1
xz-tools=5.8.1=hb9d3cd8_1
zeromq=4.3.5=h75354e8_4
zipp=3.15.0=pyhd8ed1ab_0
